% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-heat-circle.R
\name{geom_heat_circle}
\alias{geom_heat_circle}
\title{Heatcircle}
\usage{
geom_heat_circle(
  outside,
  outside_name = NULL,
  outside_colors = c("#FED7D8", "#FE8C91", "#F5636B", "#E72D3F", "#C20824"),
  inside,
  inside_name = NULL,
  inside_colors = c("gray100", "gray85", "gray50", "gray35", "gray0"),
  r = 3,
  ...
)
}
\arguments{
\item{outside}{The column name for the outside portion of heatcircle.}

\item{outside_name}{The label name (in quotes) for the legend of the outside
rendering. Default is \code{NULL}.}

\item{outside_colors}{A color vector, usually as hex codes.}

\item{inside}{The column name for the inside portion of heatcircle.}

\item{inside_name}{The label name (in quotes) for the legend of the inside
rendering. Default is \code{NULL}.}

\item{inside_colors}{A color vector, usually as hex codes.}

\item{r}{The value that controls how large of the inside portion with respect
to the outside one. When \code{r} is larger, the inside get smaller. Default
value is 3.}

\item{...}{\code{...} accepts any arguments \code{scale_fill_gradientn()} has
.}
}
\value{
A heatcircle (modified heatmap) with nested circles.
}
\description{
The heatcircle geom is used to create the two concentric circles that use
luminance to show the values from two sources on the same plot.
}
\examples{

# heatcircle with categorical variables only

data <- data.frame(x = rep(c("a", "b", "c"), 3),
                   y = rep(c("d", "e", "f"), 3),
                   outside_values = rep(c(1,5,7),3),
                   inside_values = rep(c(2,3,4),3))

ggplot(data, aes(x,y)) + geom_heat_circle(outside = outside_values,
                                          inside = inside_values)


# Making the inside smaller by setting r to be larger.

ggplot(data, aes(x,y)) + geom_heat_circle(outside = outside_values,
                                          inside = inside_values,
                                          r = 5)


# heatcircle with numeric variables only

data <- data.frame(x = rep(c(1, 2, 3), 3),
                   y = rep(c(1, 2, 3), 3),
                   outside_values = rep(c(1,5,7),3),
                   inside_values = rep(c(2,3,4),3))

ggplot(data, aes(x,y)) + geom_heat_circle(outside = outside_values,
                                          inside = inside_values)


# heatcircle with a mixure of numeric and categorical variables

data <- data.frame(x = rep(c("a", "b", "c"), 3),
                   y = rep(c(1, 2, 3), 3),
                   outside_values = rep(c(1,5,7),3),
                   inside_values = rep(c(2,3,4),3))

ggplot(data, aes(x,y)) + geom_heat_circle(outside = outside_values,
                                          inside = inside_values)

}
