% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pitts_emojis}
\alias{pitts_emojis}
\title{Popular Emojis}
\format{
An object of class \code{character} of length 270.
}
\usage{
pitts_emojis
}
\description{
The most popular emoji of a given week in a given category from the Meltwater tweet sample
They can be rendered by using \code{"richtext"} with \code{annotate()}.
}
\keyword{datasets}
