% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-options.R
\name{theme_heat}
\alias{theme_heat}
\title{Plot Themes}
\usage{
theme_heat(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\value{
Adding a heat theme to all plots generated by using the ggDoubleHeat
package.
}
\description{
Plot Themes
}
