% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-heat-circle.R
\name{geom_heat_circle}
\alias{geom_heat_circle}
\title{Heatcircle}
\usage{
geom_heat_circle(
  outside,
  outside_name = rlang::expr({     {         outside     } }),
  outside_colors = c("#FED7D8", "#FE8C91", "#F5636B", "#E72D3F", "#C20824"),
  inside,
  inside_name = rlang::expr({     {         inside     } }),
  inside_colors = c("gray100", "gray85", "gray50", "gray35", "gray0"),
  r = 3,
  ...
)
}
\arguments{
\item{outside}{The column name for the outside portion of heatcircle.}

\item{outside_name}{The label name (in quotes) for the legend of the outside rendering. Default
is the column name for the outside.}

\item{outside_colors}{A color vector, usually as hexcodes.}

\item{inside}{The column name for the inside portion of heatcircle.}

\item{inside_name}{The label name (in quotes) for the legend of the inside rendering. Default
is the column name for the inside.}

\item{inside_colors}{A color vector, usually as hexcodes.}

\item{r}{The value that controls how large of the inside portion with respect to the outside one.
When r is larger, the inside get smaller. Default value is 3.}

\item{...}{}
}
\value{

}
\description{
Heatcircle
}
