% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pitts_tg}
\alias{pitts_tg}
\title{The 30-week Pittsburgh Covid related incidence rates from Google searches and Tweets}
\format{

}
\source{
Google is processed from Google Health API, and Twitter
from Meltwater, a Twitter vendor.
}
\usage{
pitts_tg
}
\description{
A dataset containing the incidence rates of Covid related categories
incidence rates from week 1 starting from June 1 2020 to week 30
that ended in the last Sunday of the year in Pittsburgh Metropolitan
Statistical Area (MSA). The data columns are as follows:
}
\keyword{datasets}
